Disable fast sigmoid since it causes divergence (#48623)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48623

The error introduced by fast sigmoid/tanh seems to accumulate in a way that's detectable in a macro-benchmark (unfortunately I don't have the model demonstrating it in a format that can be publically committed).
ghstack-source-id: 117496822

Test Plan: Tbh not sure how to test this since I'm not super well-versed in numerics.  I can verify it fixes a model divergence locally.

Reviewed By: navahgar

Differential Revision: D25230376

fbshipit-source-id: c404a0439f190359b72ad65b3f42369c53cae340