[pt][quant] Remove contiguous calls in qembeddingbag (#48993)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48993

I don't see any reasons that we need to call contiguous on the embedding tables. They should not exist in the first place. The indices and lengths/offsets are actually generated in the model, but they're most likely generated by SigridTransform -> ClipRanges -> GatherRanges -> SigridHash (sometimes) and none of these ops produce non-contiguous tensors. It should be fine to enforce tensor.is_contiguous().

Reviewed By: radkris-git

Differential Revision: D25266756

fbshipit-source-id: f15ecb67281c9ef0c7ac6637f439e538e77e30a2